Refactors convention plugins Removes lint and detekt plugins
