--- build-logic/convention/src/main/kotlin/kovp/trainhard/convention/utils/AndroidLintConfiguration.kt
@@ -1,28 +1,0 @@
-package kovp.trainhard.convention.utils
-
-import com.android.build.api.dsl.CommonExtension
-
-private val ignoreRules = listOf(
-    "Overdraw",
-    "VectorPath",
-    "VectorRaster",
-    "SetTextI18n",
-    "AlwaysShowAction",
-    "ClickableViewAccessibility",
-    "IconLocation",
-    "LockedOrientationActivity",
-    "DataExtractionRules",
-    "UnusedResources",
-)
-
-fun CommonExtension<*, *, *, *, *, *>.configureAndroidLint() = lint {
-    abortOnError = false
-    warningsAsErrors = false
-    htmlReport = false
-    textReport = false
-    xmlReport = true
-    sarifReport = false
-    disable += ignoreRules
-    checkDependencies = true
-    showAll = true
-}
