+++ build.gradle.kts
@@ -3,7 +3,6 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.google.ksp) apply false
-    alias(libs.plugins.detekt) apply false
     alias(libs.plugins.composeCompiler) apply false
     alias(libs.plugins.kotlinx.serialization) apply false
 }
