+++ builder/plugins/src/main/kotlin/kovp/trainhard/plugins/AndroidApplicationConventionPlugin.kt
@@ -1,12 +1,10 @@
-package kovp.trainhard.convention
+package kovp.trainhard.plugins
 
 import com.android.build.gradle.internal.dsl.BaseAppModuleExtension
-import kovp.trainhard.convention.consts.Config
-import kovp.trainhard.convention.consts.Plugins
-import kovp.trainhard.convention.utils.TrainProps
-import kovp.trainhard.convention.utils.configureAndroidLint
-import kovp.trainhard.convention.utils.configureKotlinAndroid
-import kovp.trainhard.convention.utils.configureLint
+import kovp.trainhard.plugins.consts.Config
+import kovp.trainhard.plugins.consts.Plugins
+import kovp.trainhard.plugins.utils.TrainProps
+import kovp.trainhard.plugins.utils.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -25,9 +23,6 @@ class AndroidApplicationConventionPlugin : Plugin<Project> {
                     val props = TrainProps.getProperties()
 
                     targetSdk = Config.targetSdk
-
-                    configureAndroidLint()
-                    configureLint(lint)
 
                     applicationId = props["applicationId"].toString().filterNot { it == '\"' }
                     versionCode = 1
