+++ builder/plugins/src/main/kotlin/kovp/trainhard/plugins/AndroidComposeConventionPlugin.kt
@@ -1,9 +1,9 @@
-package kovp.trainhard.convention
+package kovp.trainhard.plugins
 
 import com.android.build.api.dsl.ApplicationExtension
 import com.android.build.api.dsl.LibraryExtension
-import kovp.trainhard.convention.consts.Plugins
-import kovp.trainhard.convention.utils.configureAndroidCompose
+import kovp.trainhard.plugins.consts.Plugins
+import kovp.trainhard.plugins.utils.configureAndroidCompose
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.findByType
