+++ builder/plugins/src/main/kotlin/kovp/trainhard/plugins/AndroidLibraryConventionPlugin.kt
@@ -1,15 +1,13 @@
-package kovp.trainhard.convention
+package kovp.trainhard.plugins
 
 import com.android.build.gradle.LibraryExtension
-import kovp.trainhard.convention.consts.Config
-import kovp.trainhard.convention.consts.Plugins
-import kovp.trainhard.convention.utils.addDependencies
-import kovp.trainhard.convention.utils.configureDetekt
-import kovp.trainhard.convention.utils.configureKotlinAndroid
+import kovp.trainhard.plugins.consts.Config
+import kovp.trainhard.plugins.consts.Plugins
+import kovp.trainhard.plugins.utils.addDependencies
+import kovp.trainhard.plugins.utils.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.kotlin.dsl.accessors.runtime.addDependencyTo
 import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.dependencies
 import org.gradle.kotlin.dsl.getByType
@@ -23,7 +21,6 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
             pluginManager.apply {
                 Plugins.android.let(::apply)
                 Plugins.kotlinAndroid.let(::apply)
-                Plugins.detekt.let(::apply)
             }
 
             extensions.configure<LibraryExtension> {
@@ -44,8 +41,6 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
                 }
                 add("implementation", project(":core_domain"))
             }
-
-            configureDetekt()
         }
     }
 }
