+++ builder/plugins/src/main/kotlin/kovp/trainhard/plugins/consts/Config.kt
@@ -1,4 +1,4 @@
-package kovp.trainhard.convention.consts
+package kovp.trainhard.plugins.consts
 
 object Config {
     const val targetSdk = 35
