+++ builder/plugins/src/main/kotlin/kovp/trainhard/plugins/consts/Plugins.kt
@@ -1,4 +1,4 @@
-package kovp.trainhard.convention.consts
+package kovp.trainhard.plugins.consts
 
 object Plugins {
     const val application = "com.android.application"
@@ -6,6 +6,5 @@ object Plugins {
     const val java = "java-library"
     const val kotlin = "kotlin"
     const val kotlinAndroid = "org.jetbrains.kotlin.android"
-    const val detekt = "io.gitlab.arturbosch.detekt"
     const val composeCompiler = "org.jetbrains.kotlin.plugin.compose"
 }
