+++ builder/plugins/src/main/kotlin/kovp/trainhard/plugins/utils/Utils.kt
@@ -1,4 +1,4 @@
-package kovp.trainhard.convention.utils
+package kovp.trainhard.plugins.utils
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalog
